+++ Minefield/Assets/Scripts/World/Field/Structure/Attraction.cs
@@ -3,7 +3,7 @@
 public class Attraction : Structure {
 
     public Attraction(GameObject prefab, Vector3Int origoPosition, Vector3 prefabOffset, float yAngle, int areaWidth, int areaLength,
-        float moneyOwedIncreaseAmount, int satisfactionIncreaseAmount, int secondsBetweenActions, int maxQueueLength)
+        int moneyOwedIncreaseAmount, int satisfactionIncreaseAmount, int secondsBetweenActions, int maxQueueLength)
         : base(prefab, origoPosition, prefabOffset, yAngle, areaWidth, areaLength,
             moneyOwedIncreaseAmount, satisfactionIncreaseAmount, secondsBetweenActions, 
             maxQueueLength) {
