+++ Minefield/Assets/Scripts/World/Field/Structure/Bar.cs
@@ -6,7 +6,7 @@ public class Bar : Structure {
     private int decreaseThirstAmount;
 
     public Bar(GameObject prefab, Vector3Int origoPosition, Vector3 prefabOffset, float yAngle, int areaWidth, int areaLength,
-        float moneyOwedIncreaseAmount, int satisfactionIncreaseAmount, int secondsBetweenActions, int decreaseThirstAmount, int maxQueueLength)
+        int moneyOwedIncreaseAmount, int satisfactionIncreaseAmount, int secondsBetweenActions, int decreaseThirstAmount, int maxQueueLength)
         : base(prefab, origoPosition, prefabOffset, yAngle, areaWidth, areaLength,
             moneyOwedIncreaseAmount, satisfactionIncreaseAmount, secondsBetweenActions,
             maxQueueLength) {
