+++ Minefield/Assets/Scripts/World/Field/Structure/Restaurant.cs
@@ -6,7 +6,7 @@ public class Restaurant : Structure {
     private int decreaseHungerAmount;
 
     public Restaurant(GameObject prefab, Vector3Int origoPosition, Vector3 prefabOffset, float yAngle, int areaWidth, int areaLength,
-        float moneyOwedIncreaseAmount, int satisfactionIncreaseAmount, int secondsBetweenActions, int decreaseHungerAmount, int maxQueueLength)
+        int moneyOwedIncreaseAmount, int satisfactionIncreaseAmount, int secondsBetweenActions, int decreaseHungerAmount, int maxQueueLength)
         : base(prefab, origoPosition, prefabOffset, yAngle, areaWidth, areaLength,
             moneyOwedIncreaseAmount, satisfactionIncreaseAmount, secondsBetweenActions,
             maxQueueLength) {
