+++ Minefield/Assets/Scripts/World/Field/Structure/Structure.cs
@@ -7,14 +7,14 @@ public class Structure : PopulatedField {
     protected NPC npcInside;
     protected Queue<NPC> queu;
 
-    protected float moneyOwedIncreaseAmount;
+    protected int moneyOwedIncreaseAmount;
     protected int satisfactionIncreaseAmount;
     protected int secondsBetweenActions;
     protected int maxQueueLength;
     protected DateTime lastActionTime;
 
     public Structure(GameObject prefab, Vector3Int origoPosition, Vector3 prefabOffset, float yAngle, int areaWidth, int areaLength,
-        float moneyOwedIncreaseAmount, int satisfactionIncreaseAmount, int secondsBetweenActions, int maxQueueLength)
+        int moneyOwedIncreaseAmount, int satisfactionIncreaseAmount, int secondsBetweenActions, int maxQueueLength)
         : base(prefab, origoPosition, prefabOffset, yAngle, areaWidth, areaLength) {
 
         this.moneyOwedIncreaseAmount = moneyOwedIncreaseAmount;
